+++ src/imio/smartweb/core/contents/rest/search/endpoint.py
@@ -5,27 +5,44 @@ from plone.memoize import ram
 from plone.restapi.search.handler import SearchHandler
 from plone.restapi.search.utils import unflatten_dotted_dict
 from plone.restapi.services import Service
+from plone.app.layout.navigation.interfaces import INavigationRoot
 from time import time
 
 
-@ram.cache(lambda *args: time() // (5 * 60))
-def get_news_views():
+def _cache_key(func, navigation_root):
+    uid_method = getattr(navigation_root, "UID", None)
+    uid = uid_method and uid_method() or "PloneSite"
+    return (uid, time() // (5 * 60))
+
+
+@ram.cache(_cache_key)
+def get_news_views(navigation_root):
     with api.env.adopt_user(username="admin"):
         brains = api.content.find(
-            context=api.portal.get(), portal_type="imio.smartweb.NewsView"
+            context=navigation_root, portal_type="imio.smartweb.NewsView"
         )
         # Can be improved by using index
         return {b.getObject().selected_news_folder: b.getURL() for b in brains}
 
 
-@ram.cache(lambda *args: time() // (5 * 60))
-def get_events_views():
+@ram.cache(_cache_key)
+def get_events_views(navigation_root):
     with api.env.adopt_user(username="admin"):
         brains = api.content.find(
-            context=api.portal.get(), portal_type="imio.smartweb.EventsView"
+            context=navigation_root, portal_type="imio.smartweb.EventsView"
         )
         # Can be improved by using index
         return {b.getObject().selected_agenda: b.getURL() for b in brains}
+
+
+def get_navigation_root(context):
+    """Return given context navigation root"""
+    if INavigationRoot.providedBy(context):
+        return context
+    else:
+        for item in context.aq_chain:
+            if INavigationRoot.providedBy(item):
+                return item
 
 
 def get_default_view_url(view_type):
@@ -38,11 +55,11 @@ def get_default_view_url(view_type):
         return api.content.get(UID=record).absolute_url()
 
 
-def get_views_mapping():
-    news = get_news_views()
+def get_views_mapping(navigation_root):
+    news = get_news_views(navigation_root)
     news["default"] = get_default_view_url("news")
 
-    events = get_events_views()
+    events = get_events_views(navigation_root)
     events["default"] = get_default_view_url("events")
 
     # Directory is a special usecase, since there is no selected entity
@@ -91,7 +108,7 @@ class ExtendedSearchHandler(SearchHandler):
 
     def _adapt_result(self, result):
         """Transform result"""
-        mapping = get_views_mapping()
+        mapping = get_views_mapping(self._navigation_root)
         result["items"] = [self._adapt_result_url(i, mapping) for i in result["items"]]
         return result
 
@@ -108,6 +125,10 @@ class ExtendedSearchHandler(SearchHandler):
         )
         return item
 
+    @property
+    def _navigation_root(self):
+        return get_navigation_root(self.context)
+
     def _update_metadata_fields(self, original, new):
         if isinstance(original, str):
             original = [original]
@@ -115,35 +136,59 @@ class ExtendedSearchHandler(SearchHandler):
 
     def _core_query(self, core):
         """Return core specific query parameters"""
-        mapping = get_views_mapping()
+        mapping = get_views_mapping(self._navigation_root)
         parameters = {
             "news": {
                 "portal_type": ["imio.news.NewsItem"],
                 "metadata_fields": ["id", "UID", "container_uid", "has_leadimage"],
-                "selected_news_folders": [
-                    k for k in mapping["imio.news.NewsItem"].keys() if k != "default"
-                ],
+                "selected_news_folders": {
+                    "query": [
+                        k
+                        for k in mapping["imio.news.NewsItem"].keys()
+                        if k != "default"
+                    ],
+                    "operator": "or",
+                },
+                "path": "/Plone",
             },
             "events": {
                 "portal_type": ["imio.events.Event"],
                 "metadata_fields": ["id", "UID", "container_uid", "has_leadimage"],
-                "selected_agendas": [
-                    k for k in mapping["imio.events.Event"].keys() if k != "default"
-                ],
+                "selected_agendas": {
+                    "query": [
+                        k for k in mapping["imio.events.Event"].keys() if k != "default"
+                    ],
+                    "operator": "or",
+                },
+                "path": "/Plone",
             },
             "directory": {
                 "portal_type": ["imio.directory.Contact"],
                 "metadata_fields": ["id", "UID", "container_uid", "has_leadimage"],
-                "selected_entities": [
-                    k
-                    for k in mapping["imio.directory.Contact"].keys()
-                    if k != "default"
-                ],
+                "selected_entities": {
+                    "query": [
+                        k
+                        for k in mapping["imio.directory.Contact"].keys()
+                        if k != "default"
+                    ],
+                    "operator": "or",
+                },
+                "path": "/Plone",
             },
         }.get(core)
+
+        required_fields = {
+            "news": "selected_news_folders",
+            "events": "selected_agendas",
+            "directory": "selected_entities",
+        }
 
         if parameters:
             parameters["core"] = api.portal.get_registry_record(
                 "smartweb.{0}_solr_core".format(core)
             )
+            required_field = required_fields[core]
+            if not parameters.get(required_field)["query"]:
+                # This avoid having unwanted results if no view was created
+                parameters[required_field] = ["None"]
         return parameters